backend/requirements.txt
mcp[cli]
httpx
playwright
pydantic
cryptography
python-dotenv
uvicorn
mvg
fastapi
starlette
boto3
jinja2
pyyaml
